Move AsUiText.kt, UiText.kt to ui_text package
Move TextFieldState.kt, TOTPTextField.kt to text_field package

Move ThemeSwitch.kt, ThemeSwitchState.kt to theme_switch package
