+++ app/src/main/java/com/denihilhamsyah/totphub/totp/presentation/TOTPEvent.kt
@@ -1,6 +1,6 @@
 package com.denihilhamsyah.totphub.totp.presentation
 
-import com.denihilhamsyah.totphub.totp.presentation.component.UiText
+import com.denihilhamsyah.totphub.totp.presentation.component.ui_text.UiText
 
 sealed interface TOTPEvent {
     data class OnSecretAdded(val uiText: UiText) : TOTPEvent
