--- app/src/main/java/com/denihilhamsyah/totphub/totp/presentation/component/TextFieldState.kt
@@ -1,6 +1,0 @@
-package com.denihilhamsyah.totphub.totp.presentation.component
-
-data class TextFieldState(
-    val text: String = "",
-    val error: UiText? = null
-)
