+++ app/src/main/java/com/denihilhamsyah/totphub/totp/presentation/component/text_field/TextFieldState.kt
@@ -1,0 +1,8 @@
+package com.denihilhamsyah.totphub.totp.presentation.component.text_field
+
+import com.denihilhamsyah.totphub.totp.presentation.component.ui_text.UiText
+
+data class TextFieldState(
+    val text: String = "",
+    val error: UiText? = null
+)
